order execution cleanup
